This causes a core graphics error sometimes as the surface is being created on a thread other than the main thread and OS X doesn't like that.
I'm suspicious we need to move the whole thing back on to the main thread here, which is annoying: it wouldn't be surprising if retrieving the buffer and the 'present' call need to be on the main thread.

Weird thing is I think future_desyn should *always* create an extra thread so I'm not sure why it works most of the time.