Add some 'feedback' properties to the window properties structure
This messes up the design a bit as you were previously able to create this without using a 'new', and the bindings can be set externally even though they're not supposed to.

However, I currently can't think of a better alternative than this - it feels like this should be a return value when creating the canvas, but also you want to use the value when setting up the properties (and there's nowhere to put a return value)

This is a bit ugly and potentially confusing but should work reliably when it's understood.